refactor(All): domain 패키지를 lotto와 user로 분리
