+++ src/main/java/lotto/view/UserIoManager.java
@@ -1,10 +1,10 @@
 package lotto.view;
 
-import lotto.model.Bonus;
-import lotto.model.Lotto;
-import lotto.model.UserLotto;
-import lotto.model.UserLottoRank;
-import lotto.model.UserMoney;
+import lotto.model.lotto.Bonus;
+import lotto.model.lotto.Lotto;
+import lotto.model.user.UserLotto;
+import lotto.model.user.UserLottoRank;
+import lotto.model.user.UserMoney;
 
 public class UserIoManager {
     private final InputView inputView;
