+++ src/test/java/lotto/LottoTest.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import lotto.model.Lotto;
+import lotto.model.lotto.Lotto;
 import lotto.util.generator.NumberGenerator;
 import lotto.util.generator.RandomNumberGenerator;
 import org.junit.jupiter.api.DisplayName;
