+++ src/test/java/lotto/model/lotto/BonusTest.java
@@ -1,7 +1,8 @@
-package lotto.model;
+package lotto.model.lotto;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import lotto.model.lotto.Bonus;
 import org.junit.jupiter.api.Test;
 
 class BonusTest {
