+++ src/test/java/lotto/model/lotto/LottoRankTest.java
@@ -1,5 +1,6 @@
-package lotto.model;
+package lotto.model.lotto;
 
+import lotto.model.lotto.LottoRank;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
