+++ src/test/java/lotto/model/user/UserMoneyTest.java
@@ -1,8 +1,9 @@
-package lotto.model;
+package lotto.model.user;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
+import lotto.model.user.UserMoney;
 import org.junit.jupiter.api.Test;
 
 class UserMoneyTest {
